chore: router add app basename
