+++ dev/subApp/public/index.html
@@ -13,19 +13,20 @@
       }
     </style>
     <script>
-      window.addEventListener('unhandledrejection', event => {
-        event.promise.catch((e)=>{
-        if (e instanceof Error) {
-          console.log('vue app get error:', event); // 捕获后自定义处理
-        }else {
-          console.log(e);
-        }
-      })
-        console.log('vue app get error:', event);
-      }, true);
-      window.addEventListener('error', event => {
-        console.log('vue app get error:', event); // 捕获后自定义处理
-      }, true);
+      // window.addEventListener('unhandledrejection', event => {
+      //   event.promise.catch((e)=>{
+      //   if (e instanceof Error) {
+      //     console.log('vue app get error:', event); // 捕获后自定义处理
+      //     throw e;
+      //   }else {
+      //     console.log(e);
+      //   }
+      // })
+      //   console.log('vue app get error:', event);
+      // }, true);
+      // window.addEventListener('error', event => {
+      //   console.log('vue app get error:', event); // 捕获后自定义处理
+      // }, true);
     </script>
     <!-- <script type="module">
       import a from 'http://localhost:9090/module.js';
