+++ dev/vue/vue.config.js
@@ -16,7 +16,7 @@ module.exports = {
   publicPath: 'http://localhost:8000',
   configureWebpack: (config) => {
     config.output.library = `sub-app-garfish-exports`;
-    config.output.jsonpFunction = `sub-app-jsonp`;
+    config.output.jsonpFunction = `vue-app-jsonp`;
     config.output.libraryTarget = 'umd';
     config.output.globalObject = 'window';
     config.devtool = 'source-map';
